+++ src/new/Geometry/UnstructuredGrids.jl
@@ -1,13 +1,13 @@
 
 """
-    struct UnstructuredGrid{Dc,Dp,Tp,Ti,B} <: ConformingTriangulation{Dc,Dp}
+    struct UnstructuredGrid{Dc,Dp,Tp,Ti,O} <: Grid{Dc,Dp}
       node_coordinates::Vector{Point{Dp,Tp}}
       cell_nodes::Table{Ti,Int32}
       reffes::Vector{<:NodalReferenceFE{Dc}}
       cell_types::Vector{Int8}
     end
 """
-struct UnstructuredGrid{Dc,Dp,Tp,Ti,B} <: ConformingTriangulation{Dc,Dp}
+struct UnstructuredGrid{Dc,Dp,Tp,Ti,O} <: Grid{Dc,Dp}
   node_coordinates::Vector{Point{Dp,Tp}}
   cell_nodes::Table{Ti,Int32}
   reffes::Vector{<:NodalReferenceFE{Dc}}
@@ -34,10 +34,10 @@ struct UnstructuredGrid{Dc,Dp,Tp,Ti,B} <: ConformingTriangulation{Dc,Dp}
 end
 
 """ 
-    UnstructuredGrid(trian::ConformingTriangulation)
+    UnstructuredGrid(trian::Grid)
 """
-function UnstructuredGrid(trian::ConformingTriangulation)
-  @assert ConformityStyle(trian) == RegularConformity() "UnstructuredGrid constructor only for regular grids"
+function UnstructuredGrid(trian::Grid)
+  @assert is_regular(trian) "UnstructuredGrid constructor only for regular grids"
   node_coordinates = collect1d(get_node_coordinates(trian))
   cell_nodes = Table(get_cell_nodes(trian))
   reffes = get_reffes(trian)
@@ -69,7 +69,7 @@ get_cell_nodes(g::UnstructuredGrid) = g.cell_nodes
 
 Build a grid with a single cell that is the given reference FE itself
 """
-function UnstructuredGrid(reffe::NodalReferenceFE)
+function UnstructuredGrid(reffe::LagrangianRefFE)
   node_coordinates = get_node_coordinates(reffe)
   cell_nodes = Table([collect(1:num_nodes(reffe)),])
   reffes = [reffe,]
@@ -80,7 +80,8 @@ end
 # From Polytope
 
 function UnstructuredGrid(::Type{<:ReferenceFE{D}},p::Polytope{D}) where D
-  reffe = NodalReferenceFE(p)
+  order = 1
+  reffe = LagrangianRefFE(Float64,p,order)
   UnstructuredGrid(reffe)
 end
 
@@ -92,7 +93,8 @@ function UnstructuredGrid(::Type{<:ReferenceFE{d}},p::Polytope) where d
   cell_nodes = Table(get_faces(p,d,0))
   reffaces = get_reffaces(Polytope{d},p)
   cell_type = get_face_type(p,d)
-  reffes = map(NodalReferenceFE,reffaces)
+  order = 1
+  reffes = map( (f)-> LagrangianRefFE(Float64,f,order), reffaces)
   UnstructuredGrid(
     node_coordinates,
     cell_nodes,
@@ -110,7 +112,8 @@ function UnstructuredGrid(x::AbstractArray{<:Point})
   node_coords = collect1d(x)
   cell_nodes = identity_table(Int,Int32,np)
   cell_type = fill(1,np)
-  reffes = [NodalReferenceFE(VERTEX),]
+  order = 1
+  reffes = [LagrangianRefFE(Float64,VERTEX,order),]
   UnstructuredGrid(
     node_coords,
     cell_nodes,
@@ -119,43 +122,43 @@ function UnstructuredGrid(x::AbstractArray{<:Point})
 end
 
 
-# Low dim grids
-
-"""
-    UnstructuredGrid(::Type{<:ReferenceFE{d}},trian::ConformingTriangulation) where d
-"""
-function UnstructuredGrid(::Type{<:ReferenceFE{d}},trian::ConformingTriangulation) where d
-  model = UnstructuredDiscreteModel(trian)
-  UnstructuredGrid(NodalReferenceFE{d},model)
-end
-
-function generate_cell_to_faces(d, grid::UnstructuredGrid, cell_to_vertices, vertex_to_cells)
-  reffes = get_reffes(grid)
-  polytopes = map(get_polytope,reffes)
-  cell_type_to_lface_to_lvertices = map( (p)->get_faces(p,d,0), polytopes )
-  cell_to_cell_type = get_cell_type(grid)
-
-  generate_cell_to_faces(
-    cell_to_vertices,
-    cell_type_to_lface_to_lvertices,
-    cell_to_cell_type,
-    vertex_to_cells)
-
-end
-
-function generate_cell_to_vertices(grid::UnstructuredGrid)
-  if has_straight_faces(grid)
-    cell_to_vertices = get_cell_nodes(grid)
-    vertex_to_node = collect(1:num_nodes(grid))
-    node_to_vertex = vertex_to_node
-  else
-    cell_to_nodes = get_cell_nodes(grid)
-    cell_to_cell_type = get_cell_type(grid)
-    reffes = get_reffes(grid)
-    cell_type_to_lvertex_to_lnode = map(get_vertex_node, reffes)
-    cell_to_vertices, vertex_to_node, node_to_vertex = _generate_cell_to_vertices(
-      cell_to_nodes,cell_to_cell_type,cell_type_to_lvertex_to_lnode)
-  end
-  (cell_to_vertices, vertex_to_node, node_to_vertex)
-end
+## Low dim grids
+#
+#"""
+#    UnstructuredGrid(::Type{<:ReferenceFE{d}},trian::Grid) where d
+#"""
+#function UnstructuredGrid(::Type{<:ReferenceFE{d}},trian::Grid) where d
+#  model = UnstructuredDiscreteModel(trian)
+#  UnstructuredGrid(NodalReferenceFE{d},model)
+#end
+#
+#function generate_cell_to_faces(d, grid::UnstructuredGrid, cell_to_vertices, vertex_to_cells)
+#  reffes = get_reffes(grid)
+#  polytopes = map(get_polytope,reffes)
+#  cell_type_to_lface_to_lvertices = map( (p)->get_faces(p,d,0), polytopes )
+#  cell_to_cell_type = get_cell_type(grid)
+#
+#  generate_cell_to_faces(
+#    cell_to_vertices,
+#    cell_type_to_lface_to_lvertices,
+#    cell_to_cell_type,
+#    vertex_to_cells)
+#
+#end
+#
+#function generate_cell_to_vertices(grid::UnstructuredGrid)
+#  if has_straight_faces(grid)
+#    cell_to_vertices = get_cell_nodes(grid)
+#    vertex_to_node = collect(1:num_nodes(grid))
+#    node_to_vertex = vertex_to_node
+#  else
+#    cell_to_nodes = get_cell_nodes(grid)
+#    cell_to_cell_type = get_cell_type(grid)
+#    reffes = get_reffes(grid)
+#    cell_type_to_lvertex_to_lnode = map(get_vertex_node, reffes)
+#    cell_to_vertices, vertex_to_node, node_to_vertex = _generate_cell_to_vertices(
+#      cell_to_nodes,cell_to_cell_type,cell_type_to_lvertex_to_lnode)
+#  end
+#  (cell_to_vertices, vertex_to_node, node_to_vertex)
+#end
