+++ test/new/GeometryTests/runtests.jl
@@ -8,13 +8,13 @@ using Test
 
 @testset "Triangulations" begin include("TriangulationsTests.jl") end
 
-#@testset "ConformingTriangulations" begin include("ConformingTriangulationsTests.jl") end
+@testset "Grids" begin include("GridsTests.jl") end
 #
 #@testset "DiscreteModels" begin include("DiscreteModelsTests.jl") end
 #
 #@testset "FaceLabelings" begin include("FaceLabelingsTests.jl") end
 #
-#@testset "UnstructuredGrids" begin include("UnstructuredGridsTests.jl") end
+@testset "UnstructuredGrids" begin include("UnstructuredGridsTests.jl") end
 #
 #@testset "CartesianGrids" begin include("CartesianGridsTests.jl") end
 #
